Update imports to canonical paths: config/testDialsConfig.js, utils/testDialsStorage.js, utils/formatTestName.js
